requirements.txt
MetaTrader5==5.0.45
pandas==2.1.4
numpy==1.24.3
ta==0.10.2
python-dotenv==1.0.0
schedule==1.2.0
loguru==0.7.2
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
asyncio-mqtt==0.16.1
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0 